XML Parser
Description
This is a HTML/JavaScript program, which takes the URL of an XML
document containing Cable Network company information, parses the XML
file, and extracts the list of cable networks,  displaying them in a
table. The JavaScript program will be embedded in an HTML file so that
it can be executed within a browser.
Error Handling
In case of a parsing error, your program should show an alert box
indicating an error was detected.
If XML file is not valid, an alert box will show displays “Error in XML
file”; Another error condition that should be checked for is an XML
file containing no cable network companies; In addition, this program
can handle the case when the XML file does not exist.